move value of default clone image into shared constant package
