--- shared/constant/privileged_plugins.go
@@ -1,9 +1,0 @@
-package constant
-
-var PrivilegedPlugins = []string{
-	"plugins/docker",
-	"plugins/gcr",
-	"plugins/ecr",
-	"woodpeckerci/plugin-docker",
-	"woodpeckerci/plugin-docker-buildx",
-}
